mv adherence_model() > modeling.predict.py
