+++ app/models/medication.py
@@ -4,7 +4,7 @@ from datetime import timedelta
 import numpy as np
 
 from app import db
-from app.adherence import DAY_STD
+from app.utils.adherence import DAY_STD
 
 
 class Prescription(db.Model):
